backend/prophet_api/requirements.txt
fastapi
uvicorn[standard]
pymongo
prophet
pandas
python-dotenv
certifi